Migrating home and components to TS
