--- client/src/components/index.js
@@ -1,7 +1,0 @@
-import CardHeader from "./CardHeader";
-import LoginForm from "./LoginForm";
-import ProjectForm from "./ProjectForm";
-import TaskInput from "./TaskInput";
-import TaskItem from "./TaskItem";
-
-export { CardHeader, LoginForm, ProjectForm, TaskInput, TaskItem };
